+++ internal/bufferutil/deserializer.go
@@ -1,4 +1,4 @@
-package common
+package bufferutil
 
 import (
 	"bytes"
