+++ pset/utils.go
@@ -20,7 +20,7 @@ import (
 	"bytes"
 	"encoding/binary"
 	"errors"
-	"github.com/vulpemventures/go-elements/common"
+	"github.com/vulpemventures/go-elements/internal/bufferutil"
 	"io"
 	"sort"
 
@@ -35,7 +35,7 @@ import (
 // serialization (writeTxWitness encodes the bitcoin protocol encoding for a
 // transaction input's witness into w).
 func writeTxWitness(wit [][]byte) ([]byte, error) {
-	s, err := common.NewSerializer(nil)
+	s, err := bufferutil.NewSerializer(nil)
 	if err != nil {
 		return nil, err
 	}
@@ -294,7 +294,7 @@ func readTxOut(txout []byte) (*transaction.TxOutput, error) {
 	if len(txout) < 45 {
 		return nil, psbt.ErrInvalidPsbtFormat
 	}
-	d := common.NewDeserializer(bytes.NewBuffer(txout))
+	d := bufferutil.NewDeserializer(bytes.NewBuffer(txout))
 	asset, err := d.ReadElementsAsset()
 	if err != nil {
 		return nil, err
@@ -323,7 +323,7 @@ func readTxOut(txout []byte) (*transaction.TxOutput, error) {
 }
 
 func writeTxOut(txout *transaction.TxOutput) ([]byte, error) {
-	s, err := common.NewSerializer(nil)
+	s, err := bufferutil.NewSerializer(nil)
 	if err != nil {
 		return nil, err
 	}
